backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
email-validator
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]